Grupo_i_computacion/backend/requirements.txt
Flask
python-dotenv
flask-restful
flask_sqlalchemy
flask_migrate
flask_jwt_extended